fix: make landing page a server component for reliable video autoplay
The "use client" directive caused the entire page to render client-side,
meaning the <video> tag wasn't in the initial server HTML. Browsers
block autoplay when the element is injected by JS after page load.

Extracted scroll-dependent screenshot and body overflow into a small
client component. The main page is now a server component — the video
tag is in the initial HTML response and autoplay works reliably.
